Fix #1552: Ensure that errors from pdflatex show up.
Tell pdflatex to die if it hits an error, and listen to its output,
rather than piping it to null. Filter this output for error notices
and display them to the user. If it gives us an error returncode,
or the file we need doesn't exist, tell the user about this too.

Change-Id: I1de23259d1f83f54adb63d2eb66c2a94594507d3